Remove sha1 chain - should default to system roots
There are two valid chains to follow for the Nutshell certificate. One is signed by “RapidSSL SHA256 CA - G3” root, and one is signed by “Equifax Secure Certificate Authority”. The Equifax chain has a SHA1 signature algorithm in the chain. This is now bad practice and has been deprecated, however, we shipped this api with the “Equifax Secure Certificate Authority” as the CA which is breaking installs. This commit removes this second chain and relies on the (OS) built in root certificates which should contain the “RapidSSL SHA256 CA - G3” root.

- https://community.qualys.com/blogs/securitylabs/2014/09/09/sha1-deprecation-what-you-need-to-know
- http://sha1affected.com/